+++ src/main/java/sp/sample/mitac/domain/services/interfaces/IValidateService.java
@@ -3,5 +3,5 @@ package sp.sample.mitac.domain.services.interfaces;
 import sp.sample.mitac.shared.CustomException;
 
 public interface IValidateService {
-    boolean checkCRC(int[] msg) throws CustomException;
+    boolean checkCRC(byte[] msg) throws CustomException;
 }
